code/copilot/pyproject.toml
[project]
name = "copilot"
version = "0.1.0"
description = "This project demonstrates the usage of the Microosft Agents SDK for building custom AI agents."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-monitor-opentelemetry>=1.8.2",
    "fastapi[standard]>=0.123.4",
    "microsoft-agents-activity>=0.6.1",
    "microsoft-agents-authentication-msal>=0.6.1",
    "microsoft-agents-hosting-aiohttp>=0.6.1",
    "microsoft-agents-hosting-core>=0.6.1",
    "microsoft-agents-hosting-fastapi>=0.6.1",
    "openai-agents>=0.6.1",
    "opentelemetry-instrumentation-aiohttp-client>=0.59b0",
    "pydantic-settings>=2.12.0",
]
